Restructured Actuary to a more sane structure
